Case # 47:
Currently we use a shipping api called dotnet ship to do realtime shipping for fedex, ups, usps and such.  Dot net ship no longer supports UPS.  Because of this I have an old version of dot net running on one server for ups and the newest version running on another server for all other carriers.  Currently if a customer chooses to enable UPS at all then all of their calculations use the older server for realtime shipping.  This needs to be changed as now the older interface causes problems for Canada Post and USPS.  So if  customer is using UPS it should be sent to the old server only, and any remaining calculations for shipping with other providers should be sent to the new server and the 2 responses should be added together for final results.


FILE NAME = Sites/Store_Engine/include/before_payment_include.asp
LINES MODIFIED = line no. 340 to 399