Bug 33:Multiple shipment order with rewards dollars marked as free
Description:
Multiple shipment order with rewards dollars marked as free 
Previously the system was perfectly working with single shipping, but it was not compatible with multiple shipping locations. for example if a user wants to ship items on different locations then it was causing logical & functional mistakes (for example not calculating shipment charges, if payment method is free then it was assigning Free to all orders, It was not deducting proper rewards amount from sub totals , etc. etc). Now it has been modified. 

File Modified:
                store_engine/Payment.asp
Lines Added :
               Line # 9