on feedback:
If I update an attribute and don’t specify a value in the dropdown column it always goes back to true for dropdown even if it was false before.

 

Also if I add on another attribute to an already existing attribute class and don’t specify a value for the dropdown again it goes to true

 

In both of these cases it should keep the existing value for dropdown if not specified.