feedback: On the grid.
I tried to sort and cant sort on item sku or attribute class I get an error.

Make item sku the first search field in the list.

 

The grid is missing the edit link.  It should be linked to the old asp page, no need to make a new edit page for now.